Edit:admin info update, guest acc sign up, guest acc create

+++ university-magazine/backend/app/Http/Controllers/MagazineController.php
@@ -11,6 +11,9 @@ class MagazineController extends Controller
 
     public function destroyMagazine(Magazine $magazine)
     {
+        if ($magazine->id == 2) {
+            return redirect('/dashboard')->with('success', 'You cannot delete this magazine!');
+        }
         // Delete the magazine
         $magazine->delete();
